+++ Models/ArchipelagoMetadata.cs
@@ -1,0 +1,10 @@
+namespace LTTPEnhancementTools.Models;
+
+public record ArchipelagoMetadata(
+    string Server,
+    int Player,
+    string PlayerName,
+    string Game,
+    string PatchFilePath,
+    string ExpectedSfcPath
+);